requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
backtrader==1.9.76.123
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.8
cycler==0.10.0
dateparser==1.0.0
gevent==21.8.0
greenlet==1.1.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
kiwisolver==1.3.2
matplotlib==3.2.2
multidict==5.1.0
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1629092076711/work
Pillow==8.3.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==19.0.0
pyparsing==2.4.7
python-binance==1.0.12
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.21
requests==2.25.1
service-identity==21.1.0
six==1.16.0
TA-Lib @ file:///Users/runner/miniforge3/conda-bld/ta-lib_1618425517328/work
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.10.0.0
tzlocal==3.0
ujson==4.1.0
urllib3==1.26.6
websocket==0.2.1
websocket-client==1.2.1
websockets==9.1
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.4.0

numpy~=1.21.2
pandas~=1.3.2